requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# macOS Integration
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyobjc-framework-Quartz==10.0
pyobjc-framework-ApplicationServices==10.0
pyobjc-framework-Vision==10.0

# AI/LLM Integration
ollama==0.1.6
google-genai==1.0.1
openai==1.59.5
anthropic==0.42.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
